new routes, added to model
SmartGoal model now has a field for “complete” (boolean), which
defaults as true when created.

I also created a route to get goals by user_id, which will be the main
get route we actually use in the app. We don’t want users to be able to
get goals for all users, just themselves.

I also added a route for setting a goal as complete, by providing the
id of the goal

Started a google doc with current api routes:
https://docs.google.com/document/d/16MAEpKa1scsLxBQTfCrCMynH6lxVX9G5p9db
C5TQ_eE/edit?usp=sharing