Cleaned up goalList view, added view for single goal (by cicking the title on goal list), implemented auth bug fix
I added some checks on the the goalList so that we are only displaying relevant info, and no nulls.  Clicking the title of a goal from the list takes you to a new page with just that goal, where we could have more info about it, or furhter options.  Also added a fix for a bug I noticed regarding losing auth - see my issue I opened for this