Setup a new navbar
Found a navbar template and tried it out.
It's mostly being controlled by our theme, but there were a few parts that didn't look right, so I went in and edited the CSS file to match our theme's colors and fonts.
It doesn't work perfectly with the "splash screen" you get when you're logged out.  I could probably fix this by turning CSS  on/off for certain states, but I didn't want to spend any more time on it if the group didn't like it