ViewGoal view update, Readable due dates
The view goal view will now show that the goal is complete, and remove the "complete" button if the goal is already complete. This will be useful if users click on a goal from complete list on the profile.  I also used moment to format the due dates displayed on due date goals on the goal list