SmartGoals "SmartPriority" Working
SmartGoals now sort by a custom priority algorithm.  Currently each goal is displaying its priority level as part of the div panel, but I will likely change the way this is displayed.  Lower numbers = higher priority.  Negative 1 priority is an overdue due date goal.  Also, if a goal is overdue, it will display "Overdue: " before the title (and since the priority is negative, overdue stuff will always be first on the list).

I want to mess with the interface a bit more, I'd like to have separate sections to divide up the goals, such as "top priority", etc., but for now it is working.

We can tweak the numbers of the algorithm if we decide the priority ordering system isn't ideal, that won't be too hard.